requirements.txt
pandas==1.4.4
pymysql
pytest-mock
